Fixed github compile error
This cast was OK for the visual studio warning level 4 but not for the github warning level
Fixed this

+++ DigDugRemade/LivesHUDComponent.cpp
@@ -28,7 +28,7 @@ void digdug::LivesHUDComponent::Notify(const HealthComponent& health)
 	const size_t nrTextures{ GetOwner()->GetChildCount() };
 	const int healthLeft{ health.GetHealth() };
 
-	for (int i{}; i < nrTextures; ++i)
+	for (int i{}; i < static_cast<int>(nrTextures); ++i)
 	{
 		if (i < healthLeft - 1) continue;
